rename strat/branch core
