+++ sympy/rules/branch/core.py
