+++ sympy/rules/branch/tests/test_tools.py
@@ -1,4 +1,4 @@
-from sympy.rules.branch.strat import canon
+from sympy.rules.branch.tools import canon
 from sympy import Basic
